Moved progression chart into '/pages' + fixed some warnings
