+++ package.json
@@ -24,6 +24,7 @@
     "react-tooltip": "^4.2.21",
     "redux": "^4.0.5",
     "storybook": "^6.1.21",
+    "tslib": "^2.3.0",
     "web-vitals": "^1.0.1",
     "yarn": "^1.22.10"
   },
