+++ src/components/GlobalColors.js
