--- src/pages/ProgressionChecker/ProgressionChecker.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export default function ProgressionChecker() {
-  return (
-    <div>
-      THIS IS PROGRESSION CHECKER
-    </div>
-  );
-}
