+++ src/pages/ProgressionChecker/SkeletonCard.js
