+++ src/pages/ProgressionChecker/main.less
