--- src/pages/index.js
@@ -1,12 +1,0 @@
-import * from react; 
-
-const HomePage = (() => { 
-    state = null 
-    if (user choses a degree) set state => degree;
-    return ( 
-        <button> degree</button>
-        <Flowchat /> 
-        <DegreeProgression/> 
-        <
-    );
-};
